// src/app/components/interfaces/control-type.enum.ts
export enum ControlTypeEnum {
  FILE = "file",
  CHECK_BOX = "checkbox",
  RADIO = "radio",
  TEXT = "text",
  TEL = "tel",
  PASSWORD = "password",
  COLOR = "color",
  SELECT = "select",
  URL = "url",
  EMAIL = "email",
  NUMBER = "number",
  RANGE = "range",
  DATE = "date",
  DATE_TIME = "datetime",
  MONTH = "month",
  TIME = "time",
  WEEK = "week"
}